+++ Entity/UserBusinessAbstract.php
@@ -12,6 +12,7 @@ namespace c975L\UserBundle\Entity;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Validator\Constraints as Assert;
 use Symfony\Component\Validator\Context\ExecutionContextInterface;
+use Misd\PhoneNumberBundle\Validator\Constraints\PhoneNumber as AssertPhoneNumber;
 use c975L\UserBundle\Validator\Constraints as c975LUserBundleAssert;
 use c975L\UserBundle\Validator\Constraints\Siret;
 use c975L\UserBundle\Validator\Constraints\Tva;
@@ -78,4 +79,18 @@ abstract class UserBusinessAbstract extends UserAbstract
      * @ORM\Column(name="business_tva", type="string", length=13, nullable=true)
      */
     protected $businessTva;
+
+    /**
+     * @AssertPhoneNumber
+     * @ORM\Column(type="phone_number")
+     * @ORM\Column(name="business_phone", type="string", nullable=true)
+     */
+    protected $businessPhone;
+
+    /**
+     * @AssertPhoneNumber
+     * @ORM\Column(type="phone_number")
+     * @ORM\Column(name="business_fax", type="string", nullable=true)
+     */
+    protected $businessFax;
 }